external: intel_prebuilts build task
BZ: 79013

intel_prebuilts new task generates prebuilt makefiles given source code makefiles.

This task depends on all the prebuilt Android.mk that
are actually included in the build, and then skip the Android.mk
that are not used anymore.

most of the actual intelligence is made in vendor/intel/commmon/external

We need a patch in build/tasks, because common/external is not included
in the right place in the android makefile flow in order to achieve our goal
when tasks are parsed, a modules_to_install variable
is set to the actual list of installed files, so that we can filter
which Android.mk we need to generate prebuilts for.

Change-Id: I6eb0ceeb4bfcebabe8af724526760679c87dd69e
Signed-off-by: Pierre Tardy <pierre.tardy@intel.com>
Reviewed-on: http://android.intel.com:8080/86762
Reviewed-by: Alberdi, Ion <ion.alberdi@intel.com>
Reviewed-by: Monnier, Olivier <olivier.monnier@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>