droidboot: Add a command line parameter indicating image is UEFI compatible
Issue: BMDA-4

The uefi compatible recovery_uefi.fstab file is added to droidboot
and recovery images.

When fastboot commands are used, flashing the boot partition requires
the OSIP related info in the boot image. Since the boot image of the
UEFI doesn't provide this information, the fastboot command fails. To
direct the control flow with UEFI support, a command line parameter
droidboot.uefi=1 is introduced. Depending on the this parameter the
control can determine if the executed droidboot image is UEFI compatible.

Change-Id: I3eae3814517ae1c8fa1ee9a6aaf3dc32e79fcc36
Signed-off-by: Ravindranath D <ravindranathx.doddi@intel.com>