Add a script to retrieve aplogs in ROS
BZ: 153600

With KK release, there is a a new verification done when installing
a new package. It checks that there are only /cache and
/tmp partitions mounted. If not, the update fails.

We can't let the /logs partition mounted in ROS anymore, so we
mount it right before leaving ROS to flush the aplogs then reboot.
To be sure the logcat command is over before rebooting, we use this
script.

Change-Id: I54bb55fdad473dad6763d6eac249f3c330ae7ade
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>