build: publish byt-m images
The current solution to publish images creates only the pub
folder but not the product folder under it which is required.

Also the images need not be published every time. This patch
creates a target to publish the byt-m binaries.

To build & publish only the factory image
make publish_factoryscripts

To build & publish only the fastboot-usb image
make publish_fastboot-usb

To build & publish both factory and fastboot-usb image and
also publish the shell scripts and ini file
make publish_ci_bytm

Change-Id: I90ce4ad8a6425b727ef89bab825438df4dfadc08
Signed-off-by: Huzefa Kankroliwala <huzefa.nomanx.kankroliwala@intel.com>
Reviewed-on: https://android.intel.com/290262
Reviewed-by: Hicks, Leesa <leesa.hicks@intel.com>
Tested-by: Hicks, Leesa <leesa.hicks@intel.com>